+++ apps/wallet-mobile/src/Staking/PoolTransition/usePoolTransitionModal.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+import {useModal} from '../../components'
+import {PoolTransitionModal} from './PoolTransitionModal'
+import {usePoolTransition, useStrings} from './usePoolTransition'
+
+export const usePoolTransitionModal = () => {
+  const {poolTransition, isPoolRetiring, isLoading, navigateToUpdate} = usePoolTransition()
+  const {openModal} = useModal()
+  const strings = useStrings()
+  const modalHeight = 700
+
+  React.useEffect(() => {
+    if (isPoolRetiring && poolTransition) {
+      openModal(
+        strings.title,
+        <PoolTransitionModal poolTransition={poolTransition} onContinue={navigateToUpdate} />,
+        modalHeight,
+      )
+    }
+  }, [isPoolRetiring, modalHeight, navigateToUpdate, openModal, poolTransition, strings.title])
+
+  return {isLoading}
+}
